Review + Test setup for overseer health depletion
Added notes to UIController regarding errors found during implementation of the system.

Reverted change to ItemManager.cs negating the step value. The step value represents pure value of change. To retain the effect of the negation, the ScrollY Input of the input manager has had the "Negate" modifier applied to retain a universal behaviour rather than an instance of